after all sized checks is done I return the minimum registers size possible for the counter. Because the library only support 32 and 64 bit platform we only need 32 bit as a small counter.

+++ include/fast_float/ascii_number.h
@@ -338,8 +338,8 @@ parse_number_string(UC const *p, UC const *pend,
     ++p;
   }
   UC const *const end_of_integer_part = p;
-  uint16_t digit_count = uint16_t(end_of_integer_part - start_digits);
-  answer.integer = span<UC const>(start_digits, size_t(digit_count));
+  uint32_t digit_count = uint32_t(end_of_integer_part - start_digits);
+  answer.integer = span<UC const>(start_digits, digit_count);
 #ifndef FASTFLOAT_ONLY_POSITIVE_C_NUMBER_WO_INF_NAN
   FASTFLOAT_IF_CONSTEXPR17(basic_json_fmt) {
     // at least 1 digit in integer part, without leading zeros
@@ -368,7 +368,7 @@ parse_number_string(UC const *p, UC const *pend,
       ++p;
     }
     exponent = int16_t(before - p);
-    answer.fraction = span<UC const>(before, size_t(p - before));
+    answer.fraction = span<UC const>(before, uint32_t(p - before));
     digit_count -= exponent;
   }
 #ifndef FASTFLOAT_ONLY_POSITIVE_C_NUMBER_WO_INF_NAN
@@ -541,7 +541,7 @@ parse_int_string(UC const *p, UC const *pend, T &value,
     p++;
   }
 
-  uint8_t const digit_count = uint8_t(p - start_digits);
+  uint32_t const digit_count = uint32_t(p - start_digits);
 
   if (digit_count == 0) {
     if (has_leading_zeros) {
